Create base methods for QueueHostedService
